+++ CHANGELOG.md
@@ -22,6 +22,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `rtic2` feature, renamed `rtic` to `rtic1`
 - `embedded_io::Read` trait implemented for `serial`
 
+### Breaking changes
+
+- Remove `PULL` generic from `Input` mode [#512]
+  Remove `as_push_pull_output`, `as_opendrain_output` [#512]
+
+[#512]: https://github.com/stm32-rs/stm32f1xx-hal/pull/512
+
 ## [v0.11.0] - 2025-09-09
 
 ### Breaking changes
@@ -39,6 +46,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `spi::FrameSize` and `embedded-hal` 1.0 implementations
 - Backport `I2c` implementation from `f4xx-hal`
 - move `Qei` mod inside `pwm_input` mod [#516]
+- Take `&Clocks` instead of `Clocks` [#498]
 
 ### Changed